aisle 2 fix
got rid of direction mechanic and cleaned up all of the interactions.
except:
some of the dialogue is still messed up with the honey jar and gold card.
I was also not able to fix the fact that the cursor does not change when you hover over mo during hide and seek